image-gen-app/requirements.txt
streamlit>=1.34
pandas>=2.2
python-dotenv>=1.0
openai>=1.3
